App v1.0

+++ Task9/Task9/Services/CourseService.cs
@@ -1,12 +1,11 @@
-﻿using Microsoft.AspNetCore.Identity;
-using Microsoft.EntityFrameworkCore;
+﻿using Microsoft.EntityFrameworkCore;
 using Task9.Data;
 using Task9.Models;
 using Task9.Services.Interfaces;
 
 namespace Task9.Services
 {
-    public class CourseService : ICRUDService<Course>
+    public class CourseService : ICRUDService<Course>, ICourseService
     {
         private readonly Task9Context _context;
 
@@ -15,36 +14,51 @@ namespace Task9.Services
             _context = context;
         }
 
-        public async Task<List<Course>> GetAllEntitiesAsync()
+        public async Task<List<Course>> GetAllAsync()
         {
-            return await _context.Courses.Include(c => c.Groups).ToListAsync();
+            return await _context.Courses.ToListAsync();
         }
 
-        public async Task<Course> GetEntityByIdAsync(int id)
+        public async Task<Course> GetByIdAsync(int id)
         {
-            return await _context.Courses.Include(c => c.Groups).FirstOrDefaultAsync(c => c.Id == id);
+            return await _context.Courses.FindAsync(id);
         }
 
         public async Task CreateAsync(Course course)
         {
+            if (course == null)
+                throw new ArgumentNullException($"Exeption! Lost data. The object {nameof(course)} is null.");
+
             _context.Courses.Add(course);
             await _context.SaveChangesAsync();
         }
 
         public async Task UpdateAsync(Course course)
         {
-            _context.Entry(course).State = EntityState.Modified;
+            if (course == null)
+                throw new ArgumentNullException($"Exeption! Lost data. The object {nameof(course)} is null.");
+
+            _context.Courses.Update(course);
             await _context.SaveChangesAsync();
         }
 
         public async Task DeleteAsync(int id)
         {
-            Course course = await _context.Courses.Include(c => c.Groups).FirstAsync(c => c.Id == id);
+            Course course = await _context.Courses.Include(c => c.Groups).FirstOrDefaultAsync(c => c.Id == id);
+
+            if (course == null)
+                throw new InvalidOperationException($"Exeption! The course with ID={id} was not found in the context.");
+
             if (course.Groups.Count > 0)
                 throw new DbUpdateException("You cannot delete a course with groups");
 
             _context.Courses.Remove(course);
             await _context.SaveChangesAsync();
+        }
+
+        public async Task<Course> GetDetailsAsync(int id)
+        {
+            return await _context.Courses.Include(c => c.Groups).FirstOrDefaultAsync(c => c.Id == id);
         }
     }
 }
